Update README.md
This proposal suggests the allocation of a small percentage (ranging from 0.01% to 1%) of Bitcoin’s total computing power to assist in efforts to decrypt or mitigate ransomware attacks. This initiative aims to leverage the vast computational resources of the Bitcoin network to support cybersecurity efforts, enhancing the overall security landscape.